Reorganize files into new folders and add alt text to arrow icon
