+++ src/components/SearchBar.js
@@ -1,5 +1,5 @@
 import React, {Component} from 'react';
-import './SearchBar.css'
+import '../styles/SearchBar.css'
 import { Link } from 'react-router-dom';
 
 class SearchBar extends Component {
