+++ src/styles/MovieDisplay.css
