+++ src/styles/Movies.css
